Switch to using new encoder for all options
