+++ Sources/MongoSwift/Client.swift
@@ -2,7 +2,7 @@ import Foundation
 import libmongoc
 
 /// Options to use when creating a `MongoClient`.
-public struct ClientOptions: BsonEncodable {
+public struct ClientOptions {
     /// Determines whether the client should retry supported write operations
     public let retryWrites: Bool?
 
@@ -20,15 +20,10 @@ public struct ClientOptions: BsonEncodable {
         self.eventMonitoring = eventMonitoring
         self.readConcern = readConcern
     }
-
-    /// `eventMonitoring` is a field that we set on the MongoClient, and `readConcern`
-    /// is used to set a default read concern for the client after it's created, so neither
-    /// of them should be encoded with the client options.
-    public var skipFields: [String] { return ["eventMonitoring", "readConcern"] }
 }
 
 /// Options to use when listing available databases.
-public struct ListDatabasesOptions: BsonEncodable {
+public struct ListDatabasesOptions: Encodable {
     /// An optional filter for the returned databases
     public let filter: Document?
 
@@ -155,7 +150,7 @@ public class MongoClient {
      */
     public func listDatabases(options: ListDatabasesOptions? = nil) throws -> MongoCursor {
         let encoder = BsonEncoder()
-        let opts = try encoder.encode(options)
+        let opts = try encoder.encodeIfPresent(options)
         guard let cursor = mongoc_client_find_databases_with_opts(self._client, opts?.data) else {
             throw MongoError.invalidResponse()
         }
