+++ Sources/MongoSwift/ClientSession.swift
@@ -1,11 +1,11 @@
 /// Options to use when creating a ClientSession.
-public struct SessionOptions {
+public struct SessionOptions: Encodable {
     /// Specifies whether read operations should be causally ordered within the session.
     public let causalConsistency: Bool?
 }
 
 /// A session for ordering sequential operations.
-public class ClientSession: BsonEncodable {
+public class ClientSession: Encodable {
 
     /// Initializes a new client session.
     public init() {
