+++ Tests/MongoSwiftTests/CodecTests.swift
@@ -2,25 +2,10 @@
 import Nimble
 import XCTest
 
-struct TestClass: BsonEncodable {
-    let val1 = "a"
-    let val2 = 0
-    let val3 = [1, 2, [3, 4]] as [Any]
-    let val4 = TestClass2()
-    let val5 = [3, TestClass2()] as [Any]
-}
-
-struct TestClass2: BsonEncodable {
-    let x = 1
-    let y = 2
-}
-
 final class CodecTests: XCTestCase {
     static var allTests: [(String, (CodecTests) -> () throws -> Void)] {
         return [
-            ("testEncodeStructs", testEncodeStructs),
             ("testEncodeListDatabasesOptions", testEncodeListDatabasesOptions),
-            ("testNilEncodingStrategy", testNilEncodingStrategy),
             ("testStructs", testStructs),
             ("testOptionals", testOptionals),
             ("testEncodingNonBsonNumbers", testEncodingNonBsonNumbers),
@@ -32,40 +17,23 @@ final class CodecTests: XCTestCase {
         ]
     }
 
-    func testEncodeStructs() throws {
-
-        let expected: Document = [
-            "val2": 0,
-            "val3": [1, 2, [3, 4] as Document] as Document,
-            "val5": [3, ["y": 2, "x": 1] as Document] as Document,
-            "val4": ["y": 2, "x": 1] as Document,
-            "val1": "a"
-        ]
-
-        expect(try BsonEncoder().encode(TestClass())).to(equal(expected))
-    }
-
     func testEncodeListDatabasesOptions() throws {
         let options = ListDatabasesOptions(filter: Document(["a": 10]), nameOnly: true, session: ClientSession())
-        let expected: Document = ["session": Document(), "filter": ["a": 10] as Document, "nameOnly": true]
+        let expected: Document = ["filter": ["a": 10] as Document, "nameOnly": true, "session": Document()]
         expect(try BsonEncoder().encode(options)).to(equal(expected))
     }
 
-    func testNilEncodingStrategy() throws {
-        let encoderNoNils = BsonEncoder()
-        let encoderWithNils = BsonEncoder(nilStrategy: .include)
-        let emptyOptions = ListDatabasesOptions(filter: nil, nameOnly: nil, session: nil)
+    struct TestClass: Encodable {
+        let val1 = "a"
+        let val2 = 0
+        let val3 = [[1, 2], [3, 4]]
+        let val4 = TestClass2()
+        let val5 = [TestClass2()]
+    }
 
-        // Even if the object exists, don't bother encoding it if its properties are all nil
-        expect(try encoderNoNils.encode(emptyOptions)).to(beNil())
-
-        expect(try encoderWithNils.encode(emptyOptions))
-        .to(equal(["session": nil, "filter": nil, "nameOnly": nil] as Document))
-
-        let options = ListDatabasesOptions(filter: nil, nameOnly: true, session: nil)
-        expect(try encoderNoNils.encode(options)).to(equal(["nameOnly": true]))
-        expect(try encoderWithNils.encode(options))
-        .to(equal(["session": nil, "filter": nil, "nameOnly": true]))
+    struct TestClass2: Encodable {
+        let x = 1
+        let y = 2
     }
 
     struct BasicStruct: Codable, Equatable {
@@ -105,8 +73,18 @@ final class CodecTests: XCTestCase {
     /// Test encoding/decoding a variety of structs containing simple types that have 
     /// built in Codable support (strings, arrays, ints, and structs composed of them.)
     func testStructs() throws {
-        let encoder = BSONEncoder()
+        let encoder = BsonEncoder()
         let decoder = BsonDecoder()
+
+        let expected: Document = [
+            "val1": "a",
+            "val2": 0,
+            "val3": [[1, 2], [3, 4]],
+            "val4": ["x": 1, "y": 2] as Document,
+            "val5": [["x": 1, "y": 2] as Document]
+        ]
+
+        expect(try encoder.encode(TestClass())).to(equal(expected))
 
         // a basic struct 
         let basic1 = BasicStruct(int: 1, string: "hello")
@@ -148,7 +126,7 @@ final class CodecTests: XCTestCase {
 
     /// Test encoding/decoding a struct containing optional values.
     func testOptionals() throws {
-        let encoder = BSONEncoder()
+        let encoder = BsonEncoder()
         let decoder = BsonDecoder()
 
         let s1 = OptionalsStruct(int: 1, bool: true, string: "hi")
@@ -201,7 +179,7 @@ final class CodecTests: XCTestCase {
     /// Test encoding where the struct's numeric types are non-BSON
     /// and require conversions.
     func testEncodingNonBsonNumbers() throws {
-        let encoder = BSONEncoder()
+        let encoder = BsonEncoder()
 
         let s1 = Numbers(int8: 42, int16: 42, uint8: 42, uint16: 42, uint32: 42, uint64: 42, uint: 42, float: 42)
         // all should be stored as Int32s, except the float should be stored as a double
@@ -261,6 +239,7 @@ final class CodecTests: XCTestCase {
         for k in Numbers.keys {
             doc3[k] = Double(42)
         }
+
         let res3 = try decoder.decode(Numbers.self, from: doc3)
         expect(res3).to(equal(s))
 
@@ -289,7 +268,7 @@ final class CodecTests: XCTestCase {
 
     /// Test that BSON number types are encoded properly, and can be decoded from any type they are stored as
     func testBsonNumbers() throws {
-        let encoder = BSONEncoder()
+        let encoder = BsonEncoder()
         let decoder = BsonDecoder()
         // the struct we expect to get back
         let s = BsonNumbers(int: 42, int32: 42, int64: 42, double: 42)
@@ -408,7 +387,7 @@ final class CodecTests: XCTestCase {
         let res2 = try decoder.decode(AllBsonTypes.self, from: doc2)
         expect(res2).to(equal(expected))
 
-        let encoder = BSONEncoder()
+        let encoder = BsonEncoder()
         expect(try encoder.encode(expected)).to(equal(doc2))
 
         let base64 = "//8="
