+++ Tests/MongoSwiftTests/ReadWriteConcernTests.swift
@@ -191,7 +191,6 @@ final class ReadWriteConcernTests: XCTestCase {
     }
 
     func testDocuments() throws {
-        let encoder = BsonEncoder()
         let docsPath = "\(self.getSpecsPath())/read-write-concern/tests/document"
         let testFiles = try FileManager.default.contentsOfDirectory(atPath: docsPath).filter { $0.hasSuffix(".json") }
         for filename in testFiles {
